fix(storage): Fixing watchOS crash when dealing with big files (#3389)

+++ AmplifyPlugins/Storage/Sources/AWSS3StoragePlugin/Support/Internal/StoragePersistableTransferTask.swift
@@ -108,8 +108,8 @@ struct StoragePersistableMultipartUpload: Codable {
 struct StoragePersistableSubTask: Codable {
     let uploadId: UploadID
     let partNumber: PartNumber
-    let bytes: Int
-    let bytesTransferred: Int
+    let bytes: UInt64
+    let bytesTransferred: UInt64
     let taskIdentifier: TaskIdentifier? // once an UploadPart starts uploading it will have a taskIdentifier
     let eTag: String?
